+++ packages/match-deep/package.json
@@ -36,7 +36,7 @@
     "eslint-import-resolver-typescript": "^1.1.1",
     "eslint-plugin-import": "^2.18.2",
     "jest": "^24.8.0",
-    "mockyeah-tools": "^0.24.0-alpha.0",
+    "@mockyeah/tools": "^0.24.0-alpha.0",
     "prettier": "^1.18.2",
     "typescript": "^3.5.3",
     "webpack": "^4.35.3",
