+++ packages/mockyeah-cli/README.md
@@ -1,11 +1,11 @@
-# mockyeah-cli
+# @mockyeah/cli
 
 Command line utility for [mockyeah](https://github.com/mockyeah/mockyeah),
 a powerful service mocking, recording, and playback utility.
 
 <img src="https://raw.githubusercontent.com/mockyeah/mockyeah/master/packages/mockyeah-docs/book/images/logo/mockyeah-600.png" height="200" />
 
-[![npm](https://img.shields.io/npm/v/mockyeah-cli.svg)](https://www.npmjs.com/package/mockyeah-cli)
+[![npm](https://img.shields.io/npm/v/@mockyeah/cli.svg)](https://www.npmjs.com/package/@mockyeah/cli)
 
 More at **https://mockyeah.js.org/CLI/CLI.html**.
 
@@ -31,4 +31,4 @@ Usage: mockyeah [options] [command]
 
 ## License
 
-mockyeah-cli is released under the [MIT License](https://opensource.org/licenses/MIT).
+@mockyeah/cli is released under the [MIT License](https://opensource.org/licenses/MIT).
