+++ packages/mockyeah-docs/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mockyeah-docs",
+  "name": "@mockyeah/docs",
   "version": "0.0.0",
   "description": "Documentation for a powerful service mocking, recording, and playback utility.",
   "main": "index.js",
@@ -18,7 +18,7 @@
     "gitbook-cli": "^2.3.2",
     "gitbook-plugin-expandable-chapters": "^0.2.0",
     "gitbook-plugin-github": "2",
-    "mockyeah-tools": "^0.24.0-alpha.0",
+    "@mockyeah/tools": "^0.24.0-alpha.0",
     "nyc": "^11.4.1",
     "prettier": "^1.18.2"
   }
