+++ packages/mockyeah-fetch/README.md
@@ -1,14 +1,14 @@
-# mockyeah-fetch
+# @mockyeah/fetch
 
 `fetch` integration utility for [mockyeah](https://github.com/mockyeah/mockyeah),
 a powerful service mocking, recording, and playback utility.
 
 <img src="https://raw.githubusercontent.com/mockyeah/mockyeah/master/packages/mockyeah-docs/book/images/logo/mockyeah-600.png" height="200" />
 
-[![npm](https://img.shields.io/npm/v/mockyeah-fetch.svg)](https://www.npmjs.com/package/mockyeah-fetch)
+[![npm](https://img.shields.io/npm/v/@mockyeah/fetch.svg)](https://www.npmjs.com/package/@mockyeah/fetch)
 
 More at **https://mockyeah.js.org/Integration.html**.
 
 ## License
 
-mockyeah-fetch is released under the [MIT License](https://opensource.org/licenses/MIT).
+@mockyeah/fetch is released under the [MIT License](https://opensource.org/licenses/MIT).
