+++ packages/mockyeah-fetch/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mockyeah-fetch",
+  "name": "@mockyeah/fetch",
   "version": "0.24.0-alpha.0",
   "description": "fetch integration for mockyeah request mocking utility.",
   "main": "dist/main.js",
@@ -70,7 +70,7 @@
     "babel-loader": "^8.0.5",
     "isomorphic-fetch": "^2.2.1",
     "jest": "^24.8.0",
-    "mockyeah-tools": "^0.24.0-alpha.0",
+    "@mockyeah/tools": "^0.24.0-alpha.0",
     "prettier": "^1.18.2",
     "source-map-loader": "^0.2.4",
     "typescript": "^3.5.3",
