+++ packages/mockyeah-test-jest/.eslintrc.js
@@ -1,1 +1,1 @@
-module.exports = require('mockyeah-tools/.eslintrc.js');
+module.exports = require('@mockyeah/tools/.eslintrc.js');
