+++ packages/mockyeah-test-jest/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mockyeah-test-jest",
+  "name": "@mockyeah/test-jest",
   "version": "0.24.0-alpha.0",
   "description": "Jest unit test setup for mockyeah request mocking utility.",
   "main": "index.js",
@@ -49,13 +49,13 @@
   ],
   "devDependencies": {
     "jest": "^24.8.0",
-    "mockyeah-tools": "^0.24.0-alpha.0",
+    "@mockyeah/tools": "^0.24.0-alpha.0",
     "prettier": "^1.18.2"
   },
   "engines": {
     "node": ">=8.0.0"
   },
   "dependencies": {
-    "mockyeah": "^0.24.0-alpha.0"
+    "@mockyeah/server": "^0.24.0-alpha.0"
   }
 }
