+++ packages/mockyeah-test-jest/prettier.config.js
@@ -1,1 +1,1 @@
-module.exports = require('mockyeah-tools/prettier.config.js');
+module.exports = require('@mockyeah/tools/prettier.config.js');
