+++ packages/mockyeah-test-mocha/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mockyeah-test-mocha",
+  "name": "@mockyeah/test-mocha",
   "version": "0.24.0-alpha.0",
   "description": "Mocha unit test setup for mockyeah request mocking utility.",
   "main": "index.js",
@@ -49,13 +49,13 @@
   ],
   "devDependencies": {
     "mocha": "^6.1.4",
-    "mockyeah-tools": "^0.24.0-alpha.0",
+    "@mockyeah/tools": "^0.24.0-alpha.0",
     "prettier": "^1.18.2"
   },
   "engines": {
     "node": ">=8.0.0"
   },
   "dependencies": {
-    "mockyeah": "^0.24.0-alpha.0"
+    "@mockyeah/server": "^0.24.0-alpha.0"
   }
 }
