+++ packages/mockyeah-tools/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mockyeah-tools",
+  "name": "@mockyeah/tools",
   "version": "0.24.0-alpha.0",
   "description": "local package centralizing monorepo tools",
   "private": true,
