+++ packages/mockyeah/README.md
@@ -1,10 +1,10 @@
-# mockyeah
+# @mockyeah/server
 
 **A powerful service mocking, recording, and playback utility.**
 
 <img src="https://raw.githubusercontent.com/mockyeah/mockyeah/master/packages/mockyeah-docs/book/images/logo/mockyeah-600.png" height="200" />
 
-[![npm](https://img.shields.io/npm/v/mockyeah.svg)](https://www.npmjs.com/package/mockyeah)
+[![npm](https://img.shields.io/npm/v/@mockyeah/server.svg)](https://www.npmjs.com/package/@mockyeah/server)
 [![Travis](https://img.shields.io/travis/mockyeah/mockyeah.svg)](https://travis-ci.org/mockyeah/mockyeah)
 [![Coveralls github](https://img.shields.io/coveralls/github/mockyeah/mockyeah.svg)](https://coveralls.io/github/mockyeah/mockyeah)
 
@@ -12,4 +12,4 @@ More at **https://mockyeah.js.org**.
 
 ## License
 
-mockyeah is released under the [MIT License](https://opensource.org/licenses/MIT).
+@mockyeah/server is released under the [MIT License](https://opensource.org/licenses/MIT).
