+++ packages/mockyeah/app/index.js
@@ -5,7 +5,7 @@ const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 const _ = require('lodash');
 const fetch = require('isomorphic-fetch');
-const MockyeahFetch = require('mockyeah-fetch');
+const MockyeahFetch = require('@mockyeah/fetch');
 const Logger = require('./lib/Logger');
 const proxyRoute = require('./lib/proxyRoute');
 const { modifyMockForMockyeahFetch } = require('./lib/modifyMockForMockyeahFetch');
