+++ packages/mockyeah/app/makeAPI/makePlay.js
@@ -24,7 +24,7 @@ const makePlay = app => {
       app.log(['serve', 'play'], name);
 
       suite.forEach((c, i) => {
-        // TODO: Handle `suiteName` and `suiteIndex` in `mockyeah-fetch` on mount for logging.
+        // TODO: Handle `suiteName` and `suiteIndex` in `@mockyeah/fetch` on mount for logging.
         return app.locals.mockyeahFetch.all(...c, {
           suiteName: name,
           suiteIndex: i
