+++ packages/mockyeah/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mockyeah",
+  "name": "@mockyeah/server",
   "version": "0.24.0-alpha.0",
   "description": "A powerful service mocking, recording, and playback utility.",
   "main": "index.js",
@@ -65,7 +65,7 @@
     "fs-extra": "^7.0.1",
     "gitbook-plugin-github": "2.0.0",
     "mocha": "^3.1.2",
-    "mockyeah-tools": "^0.24.0-alpha.0",
+    "@mockyeah/tools": "^0.24.0-alpha.0",
     "nodemon": "^1.17.5",
     "nyc": "^11.4.1",
     "prettier": "^1.18.2",
@@ -88,7 +88,7 @@
     "lodash": "^4.17.5",
     "match-deep": "^0.24.0-alpha.0",
     "mkdirp": "^0.5.1",
-    "mockyeah-fetch": "^0.24.0-alpha.0",
+    "@mockyeah/fetch": "^0.24.0-alpha.0",
     "request": "^2.69.0",
     "tildify": "^1.1.2"
   },
