+++ packages/mockyeah/prettier.config.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
-module.exports = require('mockyeah-tools/prettier.config.js');
+module.exports = require('@mockyeah/tools/prettier.config.js');
