+++ packages/mockyeah/test/dummy/index.js
@@ -3,6 +3,6 @@
 /* eslint-disable import/no-extraneous-dependencies, node/no-extraneous-require */
 global.MOCKYEAH_ROOT = __dirname;
 
-const mockyeah = require('mockyeah');
+const mockyeah = require('@mockyeah/server');
 
 mockyeah.play('some-custom-suite');
